Add Calculator func, begin from scratch, a completely different approach
This is an attempt to use a special programming method that will allow
to incrementally and in small portions control the content of the code.
In short, all programming is carried out outside of any language
framework in a holistic and consistent form. The implementation in a
specific language is essentially just a pure translation. Everything
is in the experimental development phase, so the details are not
described.